Performance Improvements (#142)
This was done by skipping the FindSkill() operation, which searches
every trained skill (hundreds). Since the trained skills list is
never updated, it is completely unnecessary to repeat the search.
If skills are updated, it would be necessary to reacquire this list
and perform a single additional search.
Added a found cache value to TravelShortcut.
Add a CheckSkills() call when opening the window.
Until auto-detection scanning is completed this should suffice.